g.region: Add more detail into LL error messages
When setting up region in LL, this adds the actual values to the error messages in case these are computed
in the g.region call or function call itself. The value may also provide more explanation to what was actually violated.

Similar to #1140 (e0ce1558e6d4bda6b9d189a308941cc67222e366) but for the lat-lon extent cases.