grass.script: Fix collision of random names in test
The test is written in a way that it may fail (with low probability). This is because the function does not make
any promisses about that.

However, a copy-pasted size of the name left only 2 characters for the random part which increased the collision probability
significantly. This change increases the total size the name to increase the size of the random suffix.

Failing CI for a recent commit on main seems to be the first occurence of this since the test was introduced in #653 14 months ago.