Add Zenodo DOI badge to readme (#2122)
This adds a badge which shows the DOI from Zenodo. The DOI used is the "cite all versions" DOI which redirects to the latest version on Zenodo.

An alternative would be a badge (also provided by Zenodo) which shows the latest version, but this seems to be more misleading as it would show, for example, whatever is the latest release in the readme file for the main branch. The "cite all versions" DOI at least shows the version-neutral DOI, although it leads to the latest version on Zenodo (which is the latest release in terms of time).