Mention session in the missing GISRC message
This links the missing GISRC environment variable to a GRASS session in the error message. In case the error message appears without much context, it says GRASS session rather than just session. It says active session to emphasize that a finished/ended/closed session is not good enough which is likely the context when users may encounter the message (besides the case with GRASS runtime being set up, but without an open/connected mapset).

The message now says environment variable instead of just variable and leaves out the dash/hyphen for better flow of the sentence. The format of the message is issue-colon-reason.