libgis: Mention session in the missing GISRC message (#2124)
This links the missing GISRC environment variable to a GRASS session in the error message. In case the error message appears without much context, it says GRASS session rather than just session. It says active session to emphasize that a finished/ended/closed session is not good enough which is likely the context when users may encounter the message (besides the case with GRASS runtime set up, but without an open/connected mapset).

The message now says environment variable instead of just variable and leaves out the dash/hyphen for better flow of the sentence. The format of the message is issue-colon-reason.

This is highly relevant to Jupyter notebooks, but it applies to the standard _grass.script.setup.init_ function as well.