v.to.db: fix option=query (#2085)
With `option=query`, query columns can not only be columns, but also SQL expressions like `count(x)` or `avg(x)`. In these cases the check for column type fails, and it is up to the user to first create a new column of the appropriate type for storing results, before running `v.to.db`.