CI: Add check for CRLF line endings
This adds an new workflow with one new job which is check of CRLF line endings.
The idea is that this workflow (and job) can by extended to have more steps which are
other simple, custom checks like that (like those in utils or other 3rd party checks which are not proper linters).

The CRFL check used is a simple action erclu/check-crlf which is used, e.g., in MapServer.
Alternatives seem to be AODocs/check-eol which uses .gitattributes and microsoftgraph/eol-blocker
which iteracts more with the PR.