grass.gunittest: Add --config parameter
A configuration file can now be specified explicitly in the command line which is useful for providing custom configuration file, e.g., with more or less tests excluded or with differnt timeout for different platfroms or hardware setups.

Removes a useless return from the config function. Reports OSErrors from the function without traceback.