CI: Add Pylint check for selected files
This adds Pylint check job in a new workflow. It uses the current latest Pylint and latest Python. It uses Pylint to check for minimal supported Python version.

The check runs only for the grass package and most of its content is ignored. It runs for grass.jupyter and grass.benchmark.

The config file is in the python directory because that's the focus of this PR. (In the future, GUI may have a different config file, e.g., due to different naming conventions. On the other hand, in an ideal case, the whole source code will use this file or even a more strict version of this file in the future.)

The check currently fails even for the targeted files.