Add Pylint check for the main directory
This aims at running Pylint on all other files except for python/grass and gui/wxpython. Checking all files in a project is hard to do with Pylint, but pytest pylint plugin can do the job.

Many messages are disabled and complexity and similarity limits are raised. Tests in testsuite directories and some additional directories such as doc or utils are ignored. Files with a lot of issues which are unique to them are ignored explicitly.