libgis: Add function to read or read again GISRC var
The session file (GISRC file) can be read again (force read) with a double underscore function, but when GISRC environment variable changes, the old file is read because cashed value of GISRC is read. This adds a new function which forces the the update (reading) of the variable regardless of its current value similarly to the file content reading.

This also moves the 'no GISRC' error message to the dedicated reading function (from a function of opening environment/variable/session file).