grass.jupyter: Generate API doc for Jupyter package (#2284)
- Updates documentation intro to reflect that it will be officially included in v8.2.
- Adds grass.jupyter to the libpython index TOC.
- Reduces depth of the TOC to make it more clear.
- Updates copyright year for generated files.
- Generates API doc for all grass.jupyter.
- Adds Sphix documentation build to Pylint check.
- Adds all authors to init file.

The Pylint check builds the code which Sphinx needs and there is already the pytest-based Pylint check so it could potentially do some pytest-related Sphinx tasks.

The API doc is now generated for all parts of the grass package. This make it possible for the doc to say grass.jupyter rather than just jupyter. This duplicates the generated doc for grass.script and others, but in the future this should replace the per-subpackage generated doc.