grass.jupyter: Automatically determine image size based on region (#2285)
Using the aspect ratio of the computational region, automatically determine 2D rendered image dimensions. Defaults are now None and one or both values can be specified. Missing values are taken from defaults and modified to match aspect ratio of the computational region extent. Resulting images in notebooks are now smaller without white (empty) areas around the data. As a result, the image of the displayed data is often significantly bigger.