wxGUI: fix 'SpinCtrl' widget size (#1339)
In general, using sizers and not setting size explicitly works best, so I changed the SpinCtrl wrapper to (on gtk) ignore size parameter if it's smaller than the minimum you identified, so that we don't have to remove all the size parameters directly. I removed couple size settings which were greater than the minimum and would likely cause troubles. This should help with getting rid of many Gtk-critical messages.