CI: Run workflows on main and release branches, PRs, not tags (#2348)
This limits when workflows are running using the 'on' specification. Workflows now run on pushed to the main branch and the release branches and on pull requests against these branches. They no longer run on tags and feature/fix branches.

Additionally, CodeQL runs only on main and PRs to main. (Was missing for 8.0 branch already anyway.)