CI: Compile r.example and v.example (#2351)
Compiling the two example C modules was part of the release procedure. This makes it part of the CI.

For now the check is (strangely) in the Pylint workflow because it is documentation and that workflow already does a test build of Sphinx (all three need to build from source first). Other places would be a separate doc generating and checking workflow or adding these to GCC checks to ensure higher quality standard.