grass.script: Add MapsetSession for runs in other mapsets
The new MapsetSession object uses the current db/location and changes mapset in a new separate environment. A new mapset is created if requested. MapsetSession can be used as a context manager.

This combines existing create_mapset and create_environment and adds the require/create/ensure logic, one interface, and resource handling. The underlying functionality is general enough and available as a new require_create_ensure_mapset function.

The MapsetSession interface is the same as the session handle returned from grass.script.setup.init which now gets a new env property to keep the interface unified and to allow for sessionobject--agnostic code in the tests.