doc: Illustration images for r.series, r.neighbors, r.mfilter
This adds images to manual pages of r.series, r.neighbors, and r.mfilter. These modules are newly parallelized and lack images.

The images were created with d.explanation.plot and the full workflow for each image is in a notebook which plays a role of an original image file.