grass.jupyter: Allow using script as display command
This allows users to use (full or relative) path to a script instead of a display command. This allows to pass arbitrary scripts unrelated to display, but the check for first letter of the name is also just a heuristic to help the correct use and this addition just avoids constraining the user unnecessarily.