wxGUI: Report non-zero return code from modules
When a module is executed, currently, two states are reported, aborted by the user and finished. This adds a third state, finished, but with non-zero return code. This does not add much because the interpretation is unclear and the error message should be clear from the output above, but it adds some info when debuging strange issues because it prints the return code.