v.dissolve: Compute attribute aggregate statistics
In addition to geometry dissolving, compute aggregate statistics for the attribute values of dissolved features with v.db.univar.

Requires v.db.univar JSON output. v.db.select with group is used to obtain unique values of the column the dissolving is based on. Add column and update now happens for every value, column, and statistics.