r.mask.status: Check mask status through module
Instead of using low-level test of file existence with hardcoded raster path and name, use a module to retrieve status of the raster mask.

The new module r.mask.status reports presence or absence of the 2D raster mask and provides additional details about the mask.

The PR requires additional changes in the library which are not yet included.