grass.jupyter: Center notebooks around the package (#2408)
Previously, the primary notebook was the one without grass.jupyter. Now the shorter grass.jupyter is the primary one
and the original notebook is for scripting use cases only. The notebooks are still similar, so one can compare the
differences, but there is no need to keep this for future versions as the aim of having the comparison is valid only
for introduction of grass.jupyter, i.e., for 8.0 and 8.2.

The text now reflects that grass.jupyter is officially released. Some additional info is included, e.g., about the mini grant
which will become less relevant in the next versions and may be removed or significantly revised.

Other files, such as readme, now point to the notebook which is using grass.jupyter.