init: Add mapset, CRS to no CRS needed error
Although -ec, -e -c, and -c -e work as expected when creating a new mapset, -ce leads to e being interpreted as CRS. This is a correct and common behavior which cannot be disabled in argparse. The new error message includes the mapset name and the parsed CRS to cover the -ce case while remaining focused on the location versus mapset case.