init: Remove automatic vacuum
This removes automatic vacuum of SQLite database, i.e., reverts the addition of automatic vacuum when a session ends.

Doing vacuum can take long time which the current implementation hides by running the vacuum process in the background. The typical time is much longer than the 0.1 sec suggested by the sleep call.

Doing vacuum only sometimes would be less predictable for the users and doing it only for large database would mean that it will take really long time (with a lot of disk or memory usage) when it actually happens suggesting that that's something user may want to trigger manually.

Notably, SQLite doesn't do the vacuums automatically, although PostgreSQL has autovacuum on by default.

This is an alternative to #2377 which uses VACUUM for large, modified files only.