lib/parser: Apply standards to generated script
The --script now generates code which complies with Black, Flake8 (except long lines), and Pylint (except line-too-long and unused-variable).

The practice of calling main in sys.exit() and return 0 was removed (usually not needed in this context). Parser call is now in main, so the variables are not global anymore (and thus also follow the right naming practice).